feat(rule): WRONG_DATA_POINT_HEADER rule implementation
Also general rules structure was refactored

Closes #224
