+++ lib/ddf-rules/data-point-rules.js
@@ -68,14 +68,14 @@ const cacheFor = {
 
     return cache[key];
   },
-  timeKeys: dataPointDescriptor => {
-    const key = `timeKeys@${dataPointDescriptor.dataPointDetail.fileDescriptor.file}`;
+  keysByType: (dataPointDescriptor, type) => {
+    const key = `${type}Keys@${dataPointDescriptor.dataPointDetail.fileDescriptor.file}`;
 
     if (!cache[key]) {
       const conceptTypeDictionary = cacheFor.conceptTypeDictionary(dataPointDescriptor);
 
       cache[key] = Object.keys(conceptTypeDictionary)
-        .filter(conceptTypeKey => conceptTypeDictionary[conceptTypeKey] === 'time');
+        .filter(conceptTypeKey => conceptTypeDictionary[conceptTypeKey] === type);
     }
 
     return cache[key];
@@ -107,7 +107,7 @@ module.exports = {
         }));
   },
   [registry.DATA_POINT_UNEXPECTED_TIME_VALUE]: dataPointDescriptor =>
-    cacheFor.timeKeys(dataPointDescriptor)
+    cacheFor.keysByType(dataPointDescriptor, 'time')
       .filter(timeKey => !ddfTimeUtils.detectTimeType(dataPointDescriptor.dataPointRecord[timeKey]))
       .map(timeKey => new Issue(registry.DATA_POINT_UNEXPECTED_TIME_VALUE)
         .setPath(dataPointDescriptor.dataPointDetail.fileDescriptor.fullPath)
