+++ lib/ddf-rules/general-rules/shared.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = {
+  getNonDdfDirectoryDescriptors: ddfData => ddfData.ddfRoot.directoryDescriptors
+    .filter(directoryDescriptor => !directoryDescriptor.isDDF)
+};
