+++ lib/ddf-rules/registry.js
@@ -13,6 +13,7 @@ exports.INVALID_DRILL_UP = Symbol.for('INVALID_DRILL_UP');
 exports.MEASURE_VALUE_NOT_NUMERIC = Symbol.for('MEASURE_VALUE_NOT_NUMERIC');
 exports.DATA_POINT_UNEXPECTED_ENTITY_VALUE = Symbol.for('DATA_POINT_UNEXPECTED_ENTITY_VALUE');
 exports.DATA_POINT_UNEXPECTED_TIME_VALUE = Symbol.for('DATA_POINT_UNEXPECTED_TIME_VALUE');
+exports.WRONG_DATA_POINT_HEADER = Symbol.for('WRONG_DATA_POINT_HEADER');
 exports.WRONG_ENTITY_IS_HEADER = Symbol.for('WRONG_ENTITY_IS_HEADER');
 exports.WRONG_ENTITY_IS_VALUE = Symbol.for('WRONG_ENTITY_IS_VALUE');
 exports.NON_UNIQUE_ENTITY_VALUE = Symbol.for('NON_UNIQUE_ENTITY_VALUE');
@@ -45,6 +46,7 @@ exports.tags = {
   [exports.MEASURE_VALUE_NOT_NUMERIC]: [exports.WARNING_TAG, exports.DATAPOINT_TAG],
   [exports.DATA_POINT_UNEXPECTED_ENTITY_VALUE]: [exports.WAFFLE_SERVER_TAG, exports.DATAPOINT_TAG],
   [exports.DATA_POINT_UNEXPECTED_TIME_VALUE]: [exports.WAFFLE_SERVER_TAG, exports.DATAPOINT_TAG],
+  [exports.WRONG_DATA_POINT_HEADER]: [exports.WAFFLE_SERVER_TAG, exports.DATAPOINT_TAG],
   [exports.WRONG_ENTITY_IS_HEADER]: [],
   [exports.WRONG_ENTITY_IS_VALUE]: [],
   [exports.NON_UNIQUE_ENTITY_VALUE]: [exports.WAFFLE_SERVER_TAG],
@@ -69,6 +71,7 @@ exports.descriptions = {
   [exports.MEASURE_VALUE_NOT_NUMERIC]: 'Measure in data point has not numeric type',
   [exports.DATA_POINT_UNEXPECTED_ENTITY_VALUE]: 'Unexpected entity value in the data point',
   [exports.DATA_POINT_UNEXPECTED_TIME_VALUE]: 'Unexpected time value in the data point',
+  [exports.WRONG_DATA_POINT_HEADER]: 'Invalid concept in data point',
   [exports.WRONG_ENTITY_IS_HEADER]: 'Wrong "is" header',
   [exports.WRONG_ENTITY_IS_VALUE]: 'Wrong value for "is" header',
   [exports.NON_UNIQUE_ENTITY_VALUE]: 'Non unique entity value',
