+++ test/fixtures/rules-cases/wrong-data-point-header/ddf--concepts--measures.csv
@@ -1,0 +1,3 @@
+concept,concept_type,domain,name
+lat,measure,,Latitude
+lng,measure,,Longitude
