+++ test/fixtures/rules-cases/wrong-data-point-header/ddf--concepts.csv
@@ -1,0 +1,10 @@
+concept,concept_type,domain,name,drill_up
+name,string,,,
+drill_up,string,,,
+geo,entity_domain,,,
+domain,string,,Domain,
+region,entity_set,geo,Region,"[""country"",""capital""]"
+country,entity_set,geo,Country,
+capital,entity_set,geo,Capital,
+pop,measure,geo,Population,
+year,time,,year,
