+++ test/fixtures/rules-cases/wrong-data-point-header/ddf--datapoints--pop--by--country--year.csv
@@ -1,0 +1,3 @@
+country,year,pop,name
+vat,1960,100000,foo1
+vat,1970,"1,100.196",foo2
